+++ src/EventListener/GetAttributesFromDcaListener.php
@@ -9,6 +9,7 @@
 namespace HeimrichHannot\TwigTemplatesBundle\EventListener;
 
 use Contao\Config;
+use HeimrichHannot\UtilsBundle\Container\ContainerUtil;
 use HeimrichHannot\UtilsBundle\Date\DateUtil;
 use Symfony\Component\Translation\TranslatorInterface;
 
@@ -23,18 +24,27 @@ class GetAttributesFromDcaListener
      * @var DateUtil
      */
     private $dateUtil;
+    /**
+     * @var ContainerUtil
+     */
+    private $containerUtil;
 
     /**
      * RenderListener constructor.
      */
-    public function __construct(TranslatorInterface $translator, DateUtil $dateUtil)
+    public function __construct(TranslatorInterface $translator, DateUtil $dateUtil, ContainerUtil $containerUtil)
     {
         $this->translator = $translator;
         $this->dateUtil = $dateUtil;
+        $this->containerUtil = $containerUtil;
     }
 
     public function __invoke(array $attributes, $dc = null): array
     {
+        if ($this->containerUtil->isBackend()) {
+            return $attributes;
+        }
+
         // add format placeholder for accessibility reasons
         if ('text' === $attributes['type'] && isset($attributes['rgxp']) && \in_array($attributes['rgxp'], ['datim', 'date', 'time'])) {
             $attributes['placeholder'] = $this->translator->trans('huh.twig.templates.placeholder.'.$attributes['rgxp'], [